% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem.R
\name{anonymize}
\alias{anonymize}
\title{Anonimize a column in a data frame}
\usage{
anonymize(data, column, catalog = FALSE)
}
\arguments{
\item{data}{data frame}

\item{column}{number of column to make anonymous}

\item{catalog}{if TRUE returns a catalog to bind the anonymous to the original data}
}
\description{
Fast wrapper make a column in a data frame anonymous
}
\examples{
df_anon <- anonymize(largedata, 1)
}

