\name{applejuice}
\alias{applejuice}
\docType{data}
\title{Apple juice}
\description{
Apples of each of six types (Aomori--Fuji, Aomori--Jona, Aomori--Ohrin, NZ--Envy, NZ--Jazz, NZ--Fuji) 
were blended and filtered using a gauze. Fluorescence profiles of complete excitation--emission matrix 
of filtered solutions (diluted with water to 147 times) were measured using fluorescence spectroscopy machines. The sample name refers to "type--fruit number--replicate". To save space, only two apples of each types were given in the dataset. 
}
\usage{data("applejuice")}

\examples{
data(applejuice)
summary(applejuice)
}
\keyword{dataset}
