% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyEEM.R
\name{applyEEM}
\alias{applyEEM}
\title{Apply function on EEM}
\usage{
applyEEM(EEM, FUN, ...)
}
\arguments{
\item{EEM}{a list containing EEM data as created by \code{readEEM} function.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}
}
\value{
EEM class object with modified values
}
\description{
Apply function on EEM and return EEM object
}
\examples{
require(EEM)
data(applejuice)

# add 30 to all values 
applejuice_plus <- applyEEM(applejuice, "+", 30)

# apply log10 to all values
applejuice_log <- applyEEM(applejuice_plus, log10)

# find the minimum value of each sample 
lapply(applejuice, "min")
# use lapply instead of applyEEM when the returned output is not EEM class structure

}

