% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMultipleEEM.R
\name{drawMultipleEEM}
\alias{drawMultipleEEM}
\title{Draw multiple contours for EEM data}
\usage{
drawMultipleEEM(x, ggplot = TRUE, textsize = 10, ncol = NULL,
  nrow = NULL, ...)
}
\arguments{
\item{x}{a list of EEM data generated by \code{\link{readEEM}} function or 
EEMweight object generated by \code{\link{extract}}-related functions.}

\item{ggplot}{logical value whether to use ggplot plot or not. Only true is usable for now.}

\item{textsize}{(optional) text size}

\item{ncol}{An integer describing the number of columns in the layout.}

\item{nrow}{An integer describing the number of rows in the layout.}

\item{...}{(optional) arguments for \code{\link{drawEEMgg}}}
}
\value{
Multiple figures returned on the graphic device
}
\description{
Draws multiple contours for EEM data
}
\examples{
require(EEM)
data(applejuice)
drawMultipleEEM(applejuice[1:5]) # draw contour of the first to the fifth sample

}
\seealso{
\code{\link{drawEEM}},\code{\link{drawEEMgg}}
}
