% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EEM-misc.R
\name{EEM-misc}
\alias{EEM-misc}
\alias{generateColor}
\alias{generatePoint}
\title{Internal functions for EEM package}
\usage{
generatePoint(n, pch = NULL)

generateColor(n, color.palette = NULL)
}
\arguments{
\item{n}{number}

\item{pch}{Either an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{color.palette}{(optional) contour color palette. See \code{\link[grDevices]{palette}} for more details}
}
\description{
Internal functions for EEM package
}
\details{
`generatePoint` and `generateColor` are used to create point and color
vector from specified number (n) and palette.
}
\section{Functions}{
\itemize{
\item \code{generateColor}: 
}}

