% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EEM.R
\docType{package}
\name{EEM}
\alias{EEM}
\alias{EEM-package}
\title{EEM: A package for reading and preprocessing fluorescence
excitation-emission matrix}
\description{
EEM package can be used to import raw data files,
visualizing data and preparing them for multivariate
analysis
}

