% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{eem_colors}
\alias{eem_colors}
\title{color list for enelmargen.org graphs}
\format{
An object of class \code{character} of length 20.
}
\usage{
eem_colors
}
\description{
Object of type:list, to be used interactively when \code{scale_colour_eem()} fails or is unavailable as a method.
}
\examples{
# see first color:
eem_colors[1]

}
\keyword{datasets}
