% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem-EEM_converter.R
\name{eemlist2EEM}
\alias{EEM2eemlist}
\alias{eemlist2EEM}
\title{Converting between EEM and eemlist objects}
\usage{
eemlist2EEM(eemlist)
}
\arguments{
\item{eemlist}{object in eemR package}
}
\description{
Converting between EEM object (EEM package) and eemlist object (eemR package)
}

