% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{getLoading}
\alias{getReg}
\title{Extract values from other models}
\usage{
getLoading(x)

getReg(x)
}
\arguments{
\item{x}{output variable from \code{\link[stats]{prcomp}} or \code{\link[pls]{plsr}} functions}
}
\value{
A `EEMweight` list containing title and value attributes.
}
\description{
Extract values from other models
}
\examples{
data(applejuice)
applejuice_uf <- unfold(applejuice) # unfold list into matrix
result <- prcomp(applejuice_uf)
loading <- getLoading(result)
str(loading)
}

