% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factura.R
\name{factura}
\alias{factura}
\title{Read xml invoices}
\usage{
factura(file, extend = FALSE, spanish = FALSE)
}
\arguments{
\item{file}{.xml file to parse}

\item{extend}{Do you want addition information? Default = FALSE. TRUE brings RFC (Tax ID Number), Name, IVA (VAT Tax), Date, Month, Year, Total, Folio, Directory.}

\item{spanish}{Do you want column names in Spanish? Default = FALSE.}
}
\value{
Data.frame
}
\description{
Provides interface to parse information embedded in xml invoices emitted in Mexico.
SAT (tax authority) provides a data definition to be followed by all invoices.
}
\examples{
library(eem)
ggplot(data = df, aes(x=Categoria,y=Dato))+
geom_point()+
theme_eem()
}
\author{
Eduardo Flores
}
\keyword{eem}

