% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem.R
\name{gg_cal}
\alias{gg_cal}
\title{Create a Calendar with ggplot2}
\usage{
gg_cal(df, color_fill = "grey90", only_count = TRUE, title = "Calendar")
}
\arguments{
\item{df}{data.frame with "dates" and "counts" columns.}

\item{color_fill}{The color to fill date with}

\item{only_count}{Show no legend (we only have pairs of 1 or 0/NA)}

\item{title}{The title of the calendar}
}
\description{
Create a calendar with ggplot2 out of a data.frame with "dates" and "counts" columns. Uses ggplot2 and lubridate.
}
\examples{
gg_cal(df)

}

