\name{gluten}
\alias{gluten}
\docType{data}
\title{Gluten}
\description{
Pure wheat gluten and pure wheat starch were mixed at gluten ratios ranging from 0 to 100 %, in 20 % increments. The samples were set in a cell with a quartz glass window, and the samples were pressed against the glass to obtain a flat surface. This dataset contains fluorescence excitation-emission profiles of each samples with 8 replicates. To save space, only the data with gluten ratios ranging from 0 to 60 % was provided.
}
\usage{data("gluten")}
\examples{
data(gluten)
summary(gluten)
}
\references{
Kokawa, M., Fujita, K., Sugiyama, J., Tsuta, M., Shibata, M., Araki, T., & Nabetani, H. (2012). Quantification of the distributions of gluten, starch and air bubbles in dough at different mixing stages by fluorescence fingerprint imaging. Journal of Cereal Science, 55(1), 15--21.
}
\keyword{dataset}