% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.R
\name{hc_tidy_series}
\alias{hc_tidy_series}
\title{Shorcut for tidy data frame}
\usage{
hc_tidy_series(hc, data, categories, values, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{data.frame with data.}

\item{categories}{bare column name of categories (one for every series)}

\item{values}{bare column name of values}
}
\description{
Helper function to create highchart graph from tidy datasets.
}
\examples{
dat <- data.frame("id" = c(1,2,3,4,5,6), 
       "grp" = c("A","A","B","B","C","C"),
       "value" = c(10,13,9,15,11,16))

highchart() \%>\% 
 hc_chart(type = "column") \%>\% 
 hc_tidy_series(data = dat, categories = grp, values = value)
  
}
