% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem.R
\name{loadRData}
\alias{loadRData}
\title{Load an RData file interactively}
\usage{
loadRData(fileName)
}
\arguments{
\item{filename}{R object to be loaded}
}
\description{
Nifty wrapper function to load into a named object, an RData file. Taken from \url{http://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file}
}
\examples{
#Notrun
df <- loadRData("Data/data.RData")
}

