% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(EEM_uf)
}
\arguments{
\item{EEM_uf}{Unfolded EEM matrix where columns are wavelength condition and rows are samples}
}
\value{
A matrix of normalized data
}
\description{
Normalize data (area under the curve = 1)
}
\examples{
data(applejuice)
applejuice_uf <- unfold(applejuice) # unfold list into matrix
applejuice_uf_norm <- normalize(applejuice_uf) # normalize data

rowSums(abs(applejuice_uf_norm), na.rm = TRUE) # the absolute sum of each row equal to 1
}

