% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem.R
\name{order_axis}
\alias{order_axis}
\title{Interactively order columns in ggplot}
\usage{
order_axis(data, axis, column)
}
\arguments{
\item{data}{data frame}

\item{axis}{categorical axis}

\item{column}{numerical variable by which to order}
}
\description{
Order a categorical variable by the values in a numerical variable in ggplot
}
\examples{
ggplot(order_axis(df, AXIS_X, COLUMN_Y),
       aes(x = AXIS_X_o, y = COLUMN_Y))
}

