% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem.R
\name{paint_factors}
\alias{paint_factors}
\title{Interactively paint factors with discrete colors}
\usage{
paint_factors(data, column, colors, type = "fill")
}
\arguments{
\item{data}{data frame}

\item{column}{name of categorical variable}

\item{colors}{string of colors}

\item{type}{fill or color for discrete scale}
}
\description{
Paint categorical variables by a discrete scale
}
