% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEEM.R
\name{print.EEM}
\alias{print.EEM}
\title{Print EEM}
\usage{
\method{print}{EEM}(x, ...)
}
\arguments{
\item{x}{EEM class object}

\item{...}{arguments for \code{print} function}
}
\description{
Print EEM
}
\examples{
data(applejuice)
print(applejuice)

}
