% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEEM.R
\name{readEEM}
\alias{readEEM}
\title{Read raw files and return a list}
\usage{
readEEM(path = NULL)
}
\arguments{
\item{path}{path to the files or folders which contains raw files (accept a vector).}
}
\value{
\code{readEEM} returns a list containing each raw files
}
\description{
Read raw files from fluorescence spectrometer
}
\details{
The supported format is *.txt, *.csv and *.dat files from FP-8500 (JASCO), F-7000 (Hitachi Hi-tech), RF-6000 (Shimadzu) and
Aqualog (Horiba) fluorescence spectrometer. 
It is likely that outputs from different machines of the same companies are supported by this function.
Please send a word or pull request to add support for other formats.
}

