% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem.R
\name{sample_df}
\alias{sample_df}
\title{Obtain a sample of a dataframe}
\usage{
sample_df(data, n, bagging = FALSE)
}
\arguments{
\item{data}{data frame}

\item{n}{number of rows desired}

\item{bagging}{with replacement? TRUE for bagging}
}
\description{
Fast wrapper to extract a sample (with no replacement) of n rows from a data.frame
}
\examples{
small <- sample_df(largedata, 500)
}

