% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_eem}
\alias{scale_fill_eem}
\title{ggplot color scales for enelmargen.org}
\usage{
scale_fill_eem(levels)
}
\arguments{
\item{levels}{Number of diferente levels in a factor. If continous = use any number over 20.}
}
\description{
Style colors used in \emph{www.enelmargen.org} as of july-2015.
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
q <- (qplot(carat, price, data=dsamp, colour=clarity)
      + ggtitle("Diamonds Are Forever"))
q + scale_fill_eem(20)
}

