% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme.R
\name{scales_eem}
\alias{scales_eem}
\title{ggplot color scales for enelmargen.org}
\usage{
scales_eem(esniveles)
}
\value{
An object of class \code{\link{theme}}.
}
\description{
Style colors used in \emph{www.enelmargen.org} as of july-2015.
}
\note{
not ready
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
q <- (qplot(carat, price, data=dsamp, colour=clarity)
      + ggtitle("Diamonds Are Forever"))
q + scales_eem()
}

