% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem.R
\name{split_df}
\alias{split_df}
\title{Split a df into random sets}
\usage{
split_df(data, p)
}
\arguments{
\item{data}{data frame}

\item{p}{percent desired in training set (test will be 1-p)}
}
\description{
Fast wrapper to split a data frame into random training and test sets
}
\note{
returns list with two data frames
}
\examples{
sets <- split_df(largedata, 0.75)
}

