% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetEEM.R
\name{[.EEM}
\alias{[.EEM}
\title{Subset EEM list}
\usage{
\method{[}{EEM}(x, i, ...)
}
\arguments{
\item{x}{EEM class object}

\item{i}{indices specifying elements to extract}

\item{...}{arguments for \code{subset} function}
}
\description{
Subset EEM list
}
\examples{
data(applejuice)
selected <- applejuice[1-5]
}

