% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.EEM}
\alias{summary.EEM}
\title{SummarizeEEM EEM list}
\usage{
\method{summary}{EEM}(object, ...)
}
\arguments{
\item{object}{a list containing EEM data as created by \code{readEEM} function.}

\item{...}{arguments for \code{summary} function}
}
\value{
Text on console
}
\description{
Summarize by listing the sample number, names and their dimensions
}
\examples{
data(applejuice)
summary(applejuice)
}

