% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_eem}
\alias{theme_eem}
\title{ggplot color theme for enelmargen.org}
\usage{
theme_eem(font_size = 20, base_family = "Verdana", legend_bottom = TRUE,
  no_legendname = TRUE)
}
\arguments{
\item{font_size}{mmmm}

\item{base_family}{mmm}

\item{legend_bottom}{aaa}

\item{no_legendname}{aaa}
}
\value{
An object of class \code{\link{theme}}.
}
\description{
Style plots used in \emph{www.enelmargen.org} as of july-2015.
}
\details{
\code{theme_eem} implements the standard redish
background theme.
\code{theme_eem_white} implements a variant with a white
panel and background.

\emph{eem} uses "ITC Officina Sans" as its font for graphs. If
you have access to this font, you can use it with the
\pkg{extrafont} package. "Verdana" is a good substitute.
}
\examples{
library(ggplot2)
library(eem)
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
q <- (qplot(carat, price, data=dsamp, colour=clarity)
     + labs(title = "Diamonds Are Forever"))
q + theme_eem()
}

