% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEEMgg.R
\name{drawEEMgg}
\alias{drawEEMgg}
\alias{drawEEMgg.EEM}
\alias{drawEEMgg.EEMweight}
\title{Draw contour for EEM data using ggplot2}
\usage{
drawEEMgg(x, ...)

\method{drawEEMgg}{EEM}(x, n, textsize = 25, color.palette = matlab.like,
  geom = "path", nlevels = 20, exlab = "Excitation wavelength [nm]",
  emlab = "Emission wavelength [nm]", main = NULL, has_legend = TRUE,
  zlim = NULL, breaks = waiver(), flipaxis = FALSE)

\method{drawEEMgg}{EEMweight}(x, ncomp, textsize = 25,
  color.palette = matlab.like, geom = "path", nlevels = 20,
  exlab = "Excitation wavelength [nm]", emlab = "Emission wavelength [nm]",
  main = NULL, has_legend = TRUE, zlim = NULL, breaks = waiver(),
  flipaxis = FALSE)
}
\arguments{
\item{x}{a list of EEM data generated by \code{\link[EEM]{readEEM}} function or 
EEMweight object generated by \code{\link[EEM]{extract}}-related functions.}

\item{n}{sample number. The number should not exceed \code{length(EEM)}}

\item{textsize}{(optional) text size}

\item{color.palette}{(optional) contour color palette. See \code{\link[grDevices]{palette}} for more details}

\item{geom}{(optional) object to display data. Can be "polygon" or "path" but default to "path" at the moment to 
due to some bugs for "polygon".}

\item{nlevels}{(optional) number of levels used to separate range of intensity value}

\item{exlab}{(optional) excitation-axis label}

\item{emlab}{(optional) emission-axis label}

\item{main}{(optional) plot title}

\item{flipaxis}{(optional) flip axis}

\item{ncomp}{number of components}
}
\value{
A figure is returned on the graphic device
}
\description{
This function draw contour for EEM data using ggplot2. Use `ggsave` to save the contours.
}
\details{
\code{\link{drawEEM}} is faster and should be used. But since it lacks multiple plotting capacity, \code{\link{drawEEMgg}} was created to work with drawMultipleEEM.
}
\section{Methods (by class)}{
\itemize{
\item \code{EEM}: draw EEM of EEM data created by \code{\link{readEEM}} function

\item \code{EEMweight}: draw contours of the output from \code{\link[EEM]{getLoading}} and 
\code{\link[EEM]{getReg}}.
}}
\examples{
\dontrun{
require(EEM)
data(applejuice)
drawEEMgg(applejuice, 1) # draw EEM of sample no.1
drawEEMgg(applejuice, 1, color.palette = cm.colors) # draw EEM of sample no.31 with different color
drawEEMgg(applejuice, 1, nlevels = 5) # reduce nlevels

# can be combined with other ggplot2 commands
drawEEMgg(applejuice, 1) + theme_bw() 
}

}
\seealso{
\code{\link{drawEEM}},\code{\link{drawMultipleEEM}}
}

