% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{eem_colors}
\alias{eem_colors}
\title{color list for enelmargen.org graphs}
\format{\preformatted{ chr [1:20] "#A84A44" "#E47D04" "#D8A19E" "#ae8b38" "#4d7c28" ...
}}
\usage{
eem_colors
}
\description{
Object of type:list, to be used interactively when \code{scale_colour_eem()} fails or is unavailable as a method.
}
\examples{
# see first color:
eem_colors[1]
}
\keyword{datasets}

