% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findLocalMax.R
\name{findLocalMax}
\alias{findLocalMax}
\title{Find local maximum peaks}
\usage{
findLocalMax(EEM, n, threshold = 0.7)
}
\arguments{
\item{n}{sample number. The number should not exceed \code{length(EEM)}.}

\item{threshold}{threshold value. Lower the value to cover low peaks.}

\item{x}{EEM data generated by \code{\link{readEEM}} function.}
}
\value{
A dataframe of local maximum peaks
}
\description{
Find local maximum peaks in EEM data
}
\examples{
data(applejuice)
findLocalMax(applejuice, 1)
}

