% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readEEM.R
\name{readEEM}
\alias{readEEM}
\title{Read raw files and return a list}
\usage{
readEEM(pathname = NULL)
}
\arguments{
\item{pathname}{path to the files or folders which contains raw files (accept a vector).}
}
\value{
\code{readEEM} returns a list containing each raw files
}
\description{
Read raw files from fluorescence spectrometer
}
\details{
The supported format is outputs from FP-8500 (JASCO), F-7000 (Hitachi Hi-tech) and RF-6000 (Shimadzu) fluorescence spectrometer.
It is likely that outputs from different machines of the same companies are supported by this function.
Please send a word or pull request to add support for other formats.
}

