% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unfold.R
\name{unfold}
\alias{unfold}
\alias{unfold.EEM}
\title{Unfold EEM list into a matrix}
\usage{
unfold(EEM)

\method{unfold}{EEM}(EEM)
}
\arguments{
\item{EEM}{a list containing EEM data as created by \code{readEEM} function.}
}
\value{
Unfolded EEM matrix where columns are wavelength condition and rows are samples
}
\description{
Unfold EEM list into a matrix with columns as variables (wavelength conditions) and rows as samples.
}
\examples{
data(applejuice)
applejuice_uf <- unfold(applejuice) # unfold list into matrix
dim(applejuice_uf) # dimension of unfolded matrix
}

